refactor(hybriddb):Adjust the hybriddb file directory
