+++ flags.go
@@ -7,11 +7,12 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/gitsrc/IceFireDB/driver/ipfs"
-	"github.com/gitsrc/IceFireDB/driver/oss"
-	"github.com/gitsrc/IceFireDB/driver/orbitdb"
-	"github.com/gitsrc/IceFireDB/hybriddb"
 	rafthub "github.com/tidwall/uhaha"
+
+	"github.com/gitsrc/IceFireDB/driver/hybriddb"
+	"github.com/gitsrc/IceFireDB/driver/ipfs"
+	"github.com/gitsrc/IceFireDB/driver/orbitdb"
+	"github.com/gitsrc/IceFireDB/driver/oss"
 )
 
 const usage = `{{NAME}} version: {{VERSION}} ({{GITSHA}})
@@ -106,13 +107,13 @@ func confInit(conf *rafthub.Config) {
 	flag.StringVar(&conf.Auth, "auth", conf.Auth, "")
 	flag.StringVar(&conf.Advertise, "advertise", conf.Advertise, "")
 	flag.StringVar(&testNode, "t", "", "")
-	
+
 	flag.StringVar(&ipfs.IpfsDefaultConfig.EndPointConnection, "ipfs-endpoint", "", "")
 	flag.StringVar(&oss.OssDefaultConfig.EndPointConnection, "oss-endpoint", "", "")
 	flag.StringVar(&oss.OssDefaultConfig.AccessKey, "oss-ak", "", "")
-	flag.StringVar(&oss.OssDefaultConfig.Secretkey , "oss-sk", "", "")	
+	flag.StringVar(&oss.OssDefaultConfig.Secretkey, "oss-sk", "", "")
 	flag.StringVar(&orbitdb.OrbitdbDefaultConfig.Pubsubid, "pubsub-id", "", "")
-	
+
 	flag.BoolVar(&conf.TryErrors, "try-errors", conf.TryErrors, "")
 	flag.BoolVar(&conf.InitRunQuit, "init-run-quit", conf.InitRunQuit, "")
 	flag.Int64Var(&hybriddb.DefaultConfig.HotCacheSize, "hot-cache-size", hybriddb.DefaultConfig.HotCacheSize, "")
